네이밍 컨벤션 적용 및 컨트롤러 분리
네이밍 컨벤션 적용 및 컨트롤러 분리
